configure.ac: solve FIXME: Replace `main' with a function in `-lm'
AC_CHECK_LIB receives two parameters: name of the library to link and a
function for which to test. Since autoscan doesn't know about the functions
that may exist on any given library, we have to provide it with a function
that we know exists. Autotools will generate a sample program to test for
its existance.

We chose fmod, because that's the function we also use on libHello, but it
could have been any other function that lives on the math library.